Integrate rating into UserUiState and CraftsManUiState
This commit updates the `UserUiState` and `CraftsManUiState` to include a `rating` field.

The following changes were made:
- Modified `UserMapper.toUserUiState` and `MyRequestCustomerUiState.toCraftsManUiState` to accept and set a `rating` parameter.
- Updated `MoreViewModel` and `MyRequestCustomerViewModel` to fetch and provide the rating when creating `UserUiState` and `CraftsManUiState` instances.
- Adjusted `UserProfileSection` to accept a `Float` for rating.
- Refactored `BaseViewModel.tryToExecute` to accept a `CoroutineScope` in its `execute` lambda, allowing for concurrent operations like fetching user data and rating simultaneously.

+++ app/src/main/java/com/paris_2/san3a/presentation/screen/more/components/UserProfileSection.kt
@@ -30,7 +30,7 @@ import com.paris_2.san3a.presentation.shared.utils.PreviewMultiDevices
 fun UserProfileSection(
     isCraftsman: Boolean,
     name: String,
-    rating: Double,
+    rating: Float,
     review: Int,
     isVerify: Boolean,
     painter: Painter,
@@ -114,7 +114,7 @@ fun UserProfileSection(
 private fun UserProfileSectionPreview() {
     UserProfileSection(
         name = stringResource(R.string.mohammed_akkad),
-        rating = 4.7,
+        rating = 4.7f,
         review = 1127,
         onClickEdit = {},
         isVerify = true,
@@ -129,7 +129,7 @@ private fun UserProfileSectionPreview() {
 private fun UserProfileSectionWithoutVerifyPreview() {
     UserProfileSection(
         name = stringResource(R.string.mohammed_akkad),
-        rating = 4.7,
+        rating = 4.7f,
         review = 1127,
         onClickEdit = {},
         isVerify = false,
